ingestion/requirements.txt
yfinance==1.3.0
google-cloud-storage==3.10.1
pandas==3.0.2